pyproject.toml
[project]
name = "issue-tracker-api"
version = "0.1.0"
description = "Production-ready Issue Tracker CRUD API built with FastAPI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115",
    "pwdlib[argon2]>=0.3.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyjwt>=2.11.0",
    "uvicorn[standard]>=0.30",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.8",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
